+++ Clinica.Core/Requests/Profissionais/AlterarProfissionalRequest.cs
@@ -16,7 +16,7 @@ public class AlterarProfissionalRequest : BaseRequest
     public string? NumeroTelefone { get; set; }
     public string? Email { get; set; }
     public string? NumeroRegistro { get; set; }
-    public int? UnidadeFederativaId { get; set; }
+    public long? UnidadeFederativaId { get; set; }
     public UnidadeFederativa? UnidadeFederativa { get; set; }
     public string? Conselho { get; set; }
